Swagger started to work
