+++ api/swagger/swagger.yaml
@@ -3,7 +3,7 @@ info:
   version: "0.0.1"
   title: Hello World App
 # during dev, should point to your local machine
-host: localhost:10010
+host: localhost:3000
 # basePath prefixes all resource paths 
 basePath: /
 # 
@@ -20,17 +20,11 @@ produces:
 paths:
   /hello:
     # binds a127 app logic to a route
-    x-swagger-router-controller: hello_world
-    get:
+    x-swagger-router-controller: hello-world
+    post:
       description: Returns 'Hello' to the caller
       # used as the method name of the controller
-      operationId: hello
-      parameters:
-        - name: name
-          in: query
-          description: The name of the person to whom to say hello
-          required: false
-          type: string
+      operationId: sayHello
       responses:
         "200":
           description: Success
