+++ public/app/modules/task/layout/task-layout.config.js
@@ -7,8 +7,8 @@
         .config(_config);
 
     function _config($stateProvider, $urlRouterProvider) {
-        $urlRouterProvider
-            .otherwise('/tasks/');
+        // $urlRouterProvider
+        //     .otherwise('/tasks/');
 
         $stateProvider
             .state('layout.tasks', {
